[Refactoring] Move ExpandMacro and InlineMacro to their own file
